Remind me to make smaller commits; Added Antimeme, Mute, Voiceban, Strikes and Logging commands/group commands; Started on Config; Added permission checks on Hierarchy attribute; Added Prompt extension method; Started to make use of Discord Intents; Created QueueButton type; Converted Event Listeners into a partial class to tidy up the namespace; Other misc things that don't come to mind on public commands

+++ src/Db/Strike.cs
@@ -8,7 +8,7 @@ namespace Tomoe.Db
     {
         [Key]
         public int Id { get; internal set; }
-        public int LogId { get; internal set; }
+        public long LogId { get; internal set; }
         public ulong GuildId { get; internal set; }
         public ulong IssuerId { get; internal set; }
         public ulong VictimId { get; internal set; }
